refactor: move design tokens to theme
