+++ .storybook/config.js
@@ -13,7 +13,7 @@ import { withConsole } from "@storybook/addon-console"
 import { withA11y } from "@storybook/addon-a11y"
 import { action } from "@storybook/addon-actions"
 
-import { fontFamilies } from "../src/utils/presets"
+import { fontFamilies } from "../src/theme"
 import "@storybook/addon-console"
 import "storybook-chromatic"
