+++ src/components/core/IntegrationRow/IntegrationRow.js
@@ -9,7 +9,7 @@ import { Link } from "../../Link"
 import { Button } from "../Button"
 import { Heading } from "../Heading"
 import { Badge } from "../Badge"
-import { spaces, fontFamilies } from "../../../utils/presets"
+import { spaces, fontFamilies } from "../../../theme"
 import fontSizes from "../../../theme/fontSizes"
 import fonts from "../../../theme/fonts"
 import colors from "../../../theme/colors"
