+++ src/components/core/PricingCard/PricingCard.js
@@ -17,7 +17,7 @@ import colors from "../../../theme/colors"
 import fonts from "../../../theme/fonts"
 import fontSizes from "../../../theme/fontSizes"
 import fontWeights from "../../../theme/fontWeights"
-import { spaces, breakpoints, radius } from "../../../utils/presets"
+import { spaces, breakpoints, radius } from "../../../theme"
 import { Heading } from "../Heading"
 import toggleTipIcon from "./assets/toggleTipIcon.svg"
 import { capitalizeString } from "../../../utils/helpers/"
