+++ src/components/core/Switch/Switch.js
@@ -6,7 +6,7 @@ import PropTypes from "prop-types"
 import colors from "../../../theme/colors"
 import fonts from "../../../theme/fonts"
 import hiddenStyles from "../../../theme/styles/hidden"
-import { spaces } from "../../../utils/presets"
+import { spaces } from "../../../theme"
 import { Toggle } from "../Toggle"
 
 function Switch({ fieldName, fieldValue, options = {}, onChange, ...rest }) {
