+++ src/components/core/Toggle/Toggle.js
@@ -7,7 +7,7 @@ import fonts from "../../../theme/fonts"
 import colors from "../../../theme/colors"
 import tones from "../../../theme/tones"
 import hiddenStyles from "../../../theme/styles/hidden"
-import { spaces } from "../../../utils/presets"
+import { spaces } from "../../../theme"
 import { showCustomCssDeprecationMessage } from "../../../utils/maintenance/deprecationMessages"
 
 const IN_ON_POSITIONS = [`LEFT`, `RIGHT`]
