+++ src/theme/index.js
@@ -1,5 +1,6 @@
+export { default as colors } from "./colors"
+export { palette } from "./palette"
 import { keyframes } from "@emotion/core"
-export { colors, palette } from "./colors"
 export { fontFamilies, fontSizes } from "./typography"
 export { spaces } from "./spaces"
 export { breakpoints } from "./breakpoints"
