--- src/utils/____presets/breakpoints.js
@@ -1,7 +1,0 @@
-export const breakpoints = {
-  mobile: 360,
-  phablet: 550,
-  tablet: 750,
-  desktop: 1000,
-  hd: 1300,
-}
