--- src/utils/____presets/index.js
@@ -1,6 +1,0 @@
-import { keyframes } from "@emotion/core"
-export { colors, palette } from "./colors"
-export { fontFamilies, fontSizes } from "./typography"
-export { spaces } from "./spaces"
-export { breakpoints } from "./breakpoints"
-export { dimensions, radius } from "./sizes"
