--- src/utils/____presets/sizes.js
@@ -1,28 +1,0 @@
-export const radius = {
-  default: `4px`,
-  large: `8px`,
-}
-
-export const dimensions = {
-  siteHeader: {
-    height: `3.75rem`,
-  },
-  dropdown: {
-    list: {
-      width: `250px`,
-    },
-  },
-  toast: {
-    minHeight: `3rem`,
-  },
-  pageNav: {
-    minHeight: `5rem`,
-  },
-  pagePadding: {
-    mobile: `1.4rem`,
-    tablet: `2rem`,
-  },
-  buildList: {
-    indentation: `5rem`,
-  },
-}
