--- src/utils/____presets/spaces.js
@@ -1,13 +1,0 @@
-export const spaces = {
-  "3xs": `0.125rem`,
-  "2xs": `0.25rem`,
-  xs: `0.5rem`,
-  s: `0.75rem`,
-  m: `1rem`,
-  l: `1.5rem`,
-  xl: `2rem`,
-  "2xl": `2.5rem`,
-  "3xl": `3rem`,
-  "4xl": `4rem`,
-  "5xl": `6rem`,
-}
